mvp/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
PyMuPDF==1.26.4
openai==1.102.0
python-dotenv==1.1.1
google-generativeai>=0.8.0

# Simple evaluation dependencies
datasets>=2.18.0

# OCR evaluation dependencies
editdistance>=1.0.11
pandas>=1.5.0
regex>=2023.0.0
beautifulsoup4>=4.11.0
openpyxl>=3.0.0
